improved coding

+++ src/controller/user.controller.ts
@@ -36,20 +36,25 @@ export const GetUser: Function = async (req: Request, res: Response) => {
 export const UpdateUser: Function = async (req: Request, res: Response) => {
   const { name, email, role } = req.body;
   const repository = getManager().getRepository(User);
-  await repository.update(req.params.id, {
-    name,
-    email,
-    role,
-  });
   const user = <User>await repository.findOne(req.params.id);
   if (!user) {
     return res.status(204).send({ errors: { message: "User not found" } });
   }
-  return res.status(200).send(user);
+  const userData = await repository.save({
+    id: +req.params.id,
+    name,
+    email,
+    role,
+  });
+  return res.status(200).send(userData);
 };
 
 export const DeleteUser: Function = async (req: Request, res: Response) => {
   const repository = getManager().getRepository(User);
+  const user = <User>await repository.findOne(req.params.id);
+  if (!user) {
+    return res.status(204).send({message: "User not found."});
+  }
   await repository.delete(req.params.id);
-  return res.status(201).send({ message: "User deleted successfully" });
+  return res.status(200).send({ message: "User deleted successfully" });
 };
